+++ src/Z2AudioLib/Z2SoundObject.cpp
@@ -11,6 +11,7 @@
 
 #ifndef __MWERKS__
 #include <limits>
+#include <stdint.h>
 #define FLT_MAX std::numeric_limits<float>::max()
 #endif
 
@@ -487,7 +488,7 @@ void Z2SoundObjAnime::startSoundInner(const JGeometry::TVec3<f32>& pos, f32 para
     JUT_ASSERT(747, curSoundIndex_ < animation_->getNumSounds());
 
     const JAUSoundAnimationSound* animationSound = animation_->getSound(curSoundIndex_);
-    u32 user_data = (u32)animationSound;
+    uintptr_t user_data = (uintptr_t)animationSound;
     if (reverse_) {
         curSoundIndex_--;
     } else {
